VAMS/project/Generators/volcano-info-gen/pyproject.toml
[tool.poetry]
name = "volcano-info-gen"
version = "0.1.0"
description = ""
authors = ["luisgodinho <luis.godinho13@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
